Externalized claim logic to separate package
